tests/requirements.txt
requests
python-dotenv
matplotlib
tqdm